add orders page

+++ client/src/components/ui/CardsComponents/CardInfo.jsx
@@ -8,21 +8,35 @@ import {
   removeFromFavorites,
 } from "../../../redux/slices/favorites";
 
+import { selectIsAuth } from "../../../redux/slices/auth";
 import { heart, filled_heart } from "../../../assets/index";
 
 const CardInfo = ({ brand, model, cost, description, materials }) => {
   const { id } = useParams();
 
+  const isAuth = useSelector(selectIsAuth);
   const { favorites } = useSelector((state) => state.favorites);
   const dispatch = useDispatch();
 
   const toggleHeart = useCallback(() => {
-    if (favorites.includes(id)) {
-      dispatch(removeFromFavorites(id));
+    if (isAuth) {
+      if (favorites.includes(id)) {
+        dispatch(removeFromFavorites(id));
+      } else {
+        dispatch(addToFavorites(id));
+      }
     } else {
-      dispatch(addToFavorites(id));
+      alert("At first you need to login to add to favorites!");
     }
   }, [favorites, dispatch, id]);
+
+  const clickToBasket = useCallback(() => {
+    if (isAuth) {
+      return;
+    } else {
+      alert("At first you need to login to add to basket!");
+    }
+  }, []);
 
   return (
     <div className="right">
@@ -39,8 +53,8 @@ const CardInfo = ({ brand, model, cost, description, materials }) => {
             src={favorites.includes(id) ? filled_heart : heart}
           />
         </button>
-        <button className="buy">
-          <p>Buy</p>
+        <button onClick={clickToBasket} className="buy">
+          Buy
         </button>
       </div>
       <div className="description">
